Set build directory for frontend
We are making the frontend folder into a py module this means it's local static folder will
be collected by python and added to the global static.
Since we want our build output to be collected, we make webpack build to static/dist instead of build.
The /dist is so that we have our own folder in global